plans+states: Reject attempts to persist ephemeral resources
The modules runtime should always use a different strategy to keep track
of live ephemeral resource instances, and should never persist them in
the plan or state.

These checks are here just to reduce the risk that a bug in the modules
runtime could inadvertently result in an ephemeral resource instance being
persisted. This is a bit of a "defense-in-depth" strategy, because the
state and plan types all have most of their fields exported and so we can't
be sure that all modifications will go through the mutation methods.